ADEN-11851 Move Triplelift Native to separate adapter
