+++ platforms/ucp-mobile/bidders/prebid/triplelift-native.ts
@@ -1,0 +1,11 @@
+export function getTripleliftNativeContext(): object {
+	return {
+		enabled: false,
+		slots: {
+			ntv_ad: {
+				position: 'native',
+				inventoryCodes: ['Fandom_Mobile_NativeIC_320x50'],
+			},
+		},
+	};
+}
