+++ src/ad-bidders/prebid/adapters-registry.ts
@@ -18,6 +18,7 @@ import {
 	Telaria,
 	TestBidder,
 	Triplelift,
+	TripleliftNative,
 	Verizon,
 	Wikia,
 	WikiaVideo,
@@ -47,6 +48,7 @@ class AdaptersRegistry {
 		Telaria,
 		TestBidder,
 		Triplelift,
+		TripleliftNative,
 		Verizon,
 		Wikia,
 		WikiaVideo,
