+++ src/ad-bidders/prebid/adapters/index.ts
@@ -16,6 +16,7 @@ export * from './rubicon-display';
 export * from './telaria';
 export * from './test-bidder';
 export * from './triplelift';
+export * from './triplelit-native';
 export * from './verizon';
 export * from './wikia';
 export * from './wikia-video';
