+++ src/ad-bidders/prebid/adapters/triplelift.ts
@@ -1,11 +1,8 @@
-import { context } from '@ad-engine/core';
 import { PrebidAdapter } from '../prebid-adapter';
 import { PrebidAdSlotConfig } from '../prebid-models';
-import { PrebidNativeProvider } from '../native';
 
 export class Triplelift extends PrebidAdapter {
 	static bidderName = 'triplelift';
-	static nativeBidderName = 'triplelift_native';
 
 	constructor(options) {
 		super(options);
@@ -15,26 +12,10 @@ export class Triplelift extends PrebidAdapter {
 		return Triplelift.bidderName;
 	}
 
-	isNativeModeOn(): boolean {
-		return context.get('bidders.prebid.tripleliftNative.enabled');
-	}
-
 	prepareConfigForAdUnit(
 		code: string,
 		{ sizes, inventoryCodes }: PrebidAdSlotConfig,
 	): PrebidAdUnit {
-		if (context.get(`slots.${code}.isNative`)) {
-			const prebidNativeProvider = new PrebidNativeProvider();
-			if (prebidNativeProvider.isEnabled() && this.isNativeModeOn()) {
-				const nativeMediaTypes = prebidNativeProvider.getPrebidNativeConfigMediaTypes();
-				return this.prepareNativeConfig(nativeMediaTypes, code, { inventoryCodes });
-			}
-		}
-
-		return this.prepareStandardConfig(code, { sizes, inventoryCodes });
-	}
-
-	prepareStandardConfig(code: string, { sizes, inventoryCodes }: PrebidAdSlotConfig): PrebidAdUnit {
 		return {
 			code,
 			mediaTypes: {
@@ -44,25 +25,6 @@ export class Triplelift extends PrebidAdapter {
 			},
 			bids: inventoryCodes.map((inventoryCode) => ({
 				bidder: this.bidderName,
-				params: {
-					inventoryCode,
-				},
-			})),
-		};
-	}
-
-	prepareNativeConfig(
-		nativeMediaTypes: PrebidNativeMediaType,
-		code,
-		{ inventoryCodes }: PrebidAdSlotConfig,
-	): PrebidAdUnit {
-		return {
-			code,
-			mediaTypes: {
-				native: nativeMediaTypes,
-			},
-			bids: inventoryCodes.map((inventoryCode) => ({
-				bidder: Triplelift.nativeBidderName,
 				params: {
 					inventoryCode,
 				},
